+++ src/main/java/org/jabref/gui/preferences/groups/GroupsTab.java
@@ -2,14 +2,17 @@ package org.jabref.gui.preferences.groups;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.CheckBox;
+import javafx.scene.control.ComboBox;
 import javafx.scene.control.RadioButton;
 import javafx.scene.control.TextField;
 
 import org.jabref.gui.preferences.AbstractPreferenceTabView;
 import org.jabref.gui.preferences.PreferencesTab;
+import org.jabref.gui.util.ViewModelListCellFactory;
 import org.jabref.logic.l10n.Localization;
 
 import com.airhacks.afterburner.views.ViewLoader;
+import org.jabref.model.groups.GroupHierarchyType;
 
 public class GroupsTab extends AbstractPreferenceTabView<GroupsTabViewModel> implements PreferencesTab {
 
@@ -18,6 +21,8 @@ public class GroupsTab extends AbstractPreferenceTabView<GroupsTabViewModel> imp
     @FXML private CheckBox autoAssignGroup;
     @FXML private CheckBox displayGroupCount;
     @FXML private TextField keywordSeparator;
+
+    @FXML private ComboBox<GroupHierarchyType> defaultSelectionCombo;
 
     public GroupsTab() {
         ViewLoader.view(this)
@@ -32,6 +37,14 @@ public class GroupsTab extends AbstractPreferenceTabView<GroupsTabViewModel> imp
 
     public void initialize() {
         this.viewModel = new GroupsTabViewModel(dialogService, preferencesService.getGroupsPreferences());
+
+        new ViewModelListCellFactory<GroupHierarchyType>()
+                .withText(GroupHierarchyType::getFormattedName)
+                .install(defaultSelectionCombo);
+
+        defaultSelectionCombo.itemsProperty().bind(viewModel.defaultGroupsListProperty());
+        defaultSelectionCombo.valueProperty().bindBidirectional(viewModel.selectedDefaultGroup());
+
 
         groupViewModeIntersection.selectedProperty().bindBidirectional(viewModel.groupViewModeIntersectionProperty());
         groupViewModeUnion.selectedProperty().bindBidirectional(viewModel.groupViewModeUnionProperty());